clean up reconcile logic, prepare for fetching logs when updating target status
